refactor(markdown): Split monolithic markdown utilities into focused modules
- Extract markdown renderer creation and syntax rules to markdownRenderers.ts
- Extract footnote parsing and rendering to markdownFootnotes.ts
- Extract heading collection and TOC generation to markdownHeadings.ts
- Extract GitHub callout and task list post-processing to markdownPostProcessing.ts
- Update main markdown.ts to import and orchestrate modular functions
- Add parser parameter to heading functions for improved testability
